Rename + documentation
