+++ RustInterop/01-Cpp-Embed-Rust-Library/embedder.cc
