+++ RustInterop/01-Cpp-Embed-Rust-Library/src/lib.rs
