+++ RustInterop/03-Rust-Embed-Cpp-Library-Complicated/README.md
@@ -1,0 +1,5 @@
+# Rust embedding a C++ shared library (more complicated)
+
+This directory implements a more complicated version of
+[02-Rust-Embed-Cpp-Library](../02-Rust-Embed-Cpp-Library). For example, the C
+ABI is more formally defined with backwards compatibility in mind.
