+++ RustInterop/03-Rust-Embed-Cpp-Library-Complicated/src/main.rs
