--- RustInterop/03-Rust-Uses-Cpp-Library-Complicated/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "rust_embedder"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-
-# TODO(domfarolino): Delete this section.
-# Specify where to look for native libraries
-# [build]
-# rustflags = ["-C", "link-args=-Wl,-rpath,library"]
-
-# Specify where to look for native libraries
-[build]
-# Neither of these appears to have any effect, which I do not understand:
-# rustflags = ["-C", "link-args=-Wl,-rpath,library"]
-# rustflags = ["-L", "native=library"]
