--- RustInterop/03-Rust-Uses-Cpp-Library-Complicated/README.md
@@ -1,9 +1,0 @@
-# Simple C <-> Rust interop
-
-This is perhaps the most simple example of creating a C ABI implemented in Rust,
-and calling it from an existing C program. A really great resource to consult
-on this topic is:
-http://blog.asleson.org/2021/02/23/how-to-writing-a-c-shared-library-in-rust/.
-
-For more information on going the other way — calling C functions from rust — see
-https://stackoverflow.com/a/50230460/3947332.
